+++ 100-js-algo/utils/utils.js
